Add setup.py to project.
Add a setup.py for the gerritlib project. This incorporates
openstack.common.setup/version as well for versioning. This change
also adds a tox.ini file and its lists of dependencies to gerritlib.
This was not done in a separate commit as tox depends on setup.py and
setup depends on tox for gate tests.

Change-Id: I101471e836802b1c38796d8d3c570ce5f00c793e