Allow ctrl-Cing gerritlib programs.
If the GerritWatcher thread is not set as daemon, the program won't
exit when the main thread exits. This is usually undesirable.

As part of making the thread be a daemon, I've moved the ownership of
threading logic to the gerrit startWatching method, decoupling
GerritWatcher's logic from the implementation detail of being a thread
(vs a co-process etc).

Change-Id: Ib28d5bfec47d2c4c88b99e07dec50ee4dd0104bf