Add empty-commit flag when creating a project with no commits
When creating a new project, if you pass the --empty-commit
flag, a sha is created for the master branch allowing for
cloning and other subsequent actions on the repo. If the project
has no commits and is created without the --empty-commit flag,
the repo is created with no sha for the master branch creating
problems with any subsequent actions on the repo.

Change-Id: Ic233057c2503b7fd9fc234d97d71431520ea70b8